--- faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/model/ReadOpResult.java
@@ -1,3 +1,0 @@
-package org.apache.zookeeper.faaskeeper.model;
-
-public abstract class ReadOpResult {}
