+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/CreateNode.java
@@ -1,7 +1,11 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 import java.util.Map;
 import java.util.List;
+
+import org.apache.zookeeper.faaskeeper.model.Node;
+import org.apache.zookeeper.faaskeeper.model.SystemCounter;
+import org.apache.zookeeper.faaskeeper.model.Version;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.util.HashMap;
