+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/GetChildrenResult.java
@@ -1,4 +1,4 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 import java.util.List;
