+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/NodeExists.java
@@ -1,4 +1,4 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 public class NodeExists extends DirectOperation {
