+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/ReadOpResult.java
@@ -1,0 +1,3 @@
+package org.apache.zookeeper.faaskeeper.operations;
+
+public abstract class ReadOpResult {}
