+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/RegisterSession.java
@@ -1,4 +1,4 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 public class RegisterSession extends DirectOperation {
     public final String sourceAddr;
