+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/RegisterSessionResult.java
@@ -1,4 +1,4 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 public class RegisterSessionResult extends ReadOpResult {
     private String sessionID;
