+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/operations/RequestOperation.java
@@ -1,7 +1,10 @@
-package org.apache.zookeeper.faaskeeper.model;
+package org.apache.zookeeper.faaskeeper.operations;
 
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
+
+import org.apache.zookeeper.faaskeeper.model.Node;
+
 import com.fasterxml.jackson.databind.JsonNode;
 
 public abstract class RequestOperation extends Operation {
