+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/queue/WorkQueueItem.java
@@ -2,7 +2,7 @@ package org.apache.zookeeper.faaskeeper.queue;
 
 import java.util.concurrent.CompletableFuture;
 
-import org.apache.zookeeper.faaskeeper.model.Operation;
+import org.apache.zookeeper.faaskeeper.operations.Operation;
 
 public class WorkQueueItem<T> {
     public final int requestID;
