+++ faaskeeper/src/main/java/org/apache/zookeeper/faaskeeper/thread/SubmitterThread.java
@@ -14,18 +14,18 @@ import org.apache.zookeeper.faaskeeper.queue.EventQueue;
 import org.apache.zookeeper.faaskeeper.queue.WorkQueueItem;
 import org.apache.zookeeper.faaskeeper.provider.ProviderClient;
 import org.apache.zookeeper.faaskeeper.provider.NodeDoesNotExist;
-import org.apache.zookeeper.faaskeeper.model.DirectOperation;
-import org.apache.zookeeper.faaskeeper.model.GetChildren;
-import org.apache.zookeeper.faaskeeper.model.RegisterSession;
-import org.apache.zookeeper.faaskeeper.model.GetData;
 import org.apache.zookeeper.faaskeeper.model.Node;
-import org.apache.zookeeper.faaskeeper.model.NodeExists;
-import org.apache.zookeeper.faaskeeper.model.ReadOpResult;
-import org.apache.zookeeper.faaskeeper.model.RequestOperation;
-import org.apache.zookeeper.faaskeeper.model.GetDataResult;
-import org.apache.zookeeper.faaskeeper.model.GetChildrenResult;
-import org.apache.zookeeper.faaskeeper.model.RegisterSessionResult;
-import org.apache.zookeeper.faaskeeper.model.ReadExceptionResult;
+import org.apache.zookeeper.faaskeeper.operations.DirectOperation;
+import org.apache.zookeeper.faaskeeper.operations.GetChildren;
+import org.apache.zookeeper.faaskeeper.operations.GetChildrenResult;
+import org.apache.zookeeper.faaskeeper.operations.GetData;
+import org.apache.zookeeper.faaskeeper.operations.GetDataResult;
+import org.apache.zookeeper.faaskeeper.operations.NodeExists;
+import org.apache.zookeeper.faaskeeper.operations.ReadExceptionResult;
+import org.apache.zookeeper.faaskeeper.operations.ReadOpResult;
+import org.apache.zookeeper.faaskeeper.operations.RegisterSession;
+import org.apache.zookeeper.faaskeeper.operations.RegisterSessionResult;
+import org.apache.zookeeper.faaskeeper.operations.RequestOperation;
 
 public class SubmitterThread implements Runnable {
     private Future<?> future;
